requirements.txt
requests>=2.31.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
notebook>=7.0.0
pyyaml>=6.0.1
scikit-learn>=1.3.0
numpy>=1.25.0
